Fix testBasicMapping. Implement fieldQuery().
Force use of fieldQuery to create a query from query strings.
Fixes the analysis for the testBasicMapping test case.

Implementation of fieldQuery and fieldFilter.

Rename the return variable in prefixFilter (its a filter, not a query).

Added types in the request, to help finding the appropriate mapping.

Added fieldQuery on a prefix.